requirements.txt
tqdm
psycopg2
python-slugify
attrs
lxml
cssselect